educator signup helper -> loginsignup helper (merge)
