+++ COPYRIGHT
@@ -1,4 +1,4 @@
-OpenStax Accounts is Copyright Rice University, 2013-2016.
+OpenStax Accounts is Copyright Rice University, 2013-2022.
 
 Programmatic source files herein are licensed under the Affero General Public License
 version 3 (the text of which can be found in the file GNU-AGPL-3.0 or online at
