+++ app/views/signup/signup_form.html.erb
@@ -42,7 +42,7 @@
         </div>
 
         <% if selected_signup_role == :student %>
-          <%= render partial: 'social_auth/social_buttons' %>
+          <%= render partial: 'social_auth/social_auth_buttons' %>
 
           <%=
             render(
