--- coffeelint.json
@@ -1,49 +1,0 @@
-{
-  "coffeelint": {
-    "transforms": [
-      "coffee-react-transform"
-    ]
-  },
-  "indentation": {
-    "level": "error",
-    "value": 2
-  },
-  "line_endings": {
-    "value": "unix",
-    "level": "error"
-  },
-  "arrow_spacing": {
-    "level": "error"
-  },
-  "no_empty_functions": {
-    "level": "error"
-  },
-  "no_empty_param_list": {
-    "level": "error"
-  },
-  "no_interpolation_in_single_quotes": {
-    "level": "error"
-  },
-  "no_throwing_strings": {
-    "level": "error"
-  },
-  "no_unnecessary_double_quotes": {
-    "level": "ignore"
-  },
-  "no_unnecessary_fat_arrows": {
-    "level": "error"
-  },
-  "prefer_english_operator": {
-    "level": "error"
-  },
-  "space_operators": {
-    "level": "error"
-  },
-  "spacing_after_comma": {
-    "level": "error"
-  },
-  "max_line_length": {
-    "value": 135,
-    "level": "error"
-  }
-}
